requirements.txt
google-genai>=1.0.0
pinecone>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
python-docx>=1.1.0
PyPDF2>=3.0.0
flask>=3.0.0
